+++ NetBeansProjects/SistemaDeVendas/src/Interface/Cad_Produ.java
@@ -232,7 +232,7 @@ public class Cad_Produ extends javax.swing.JFrame {
             p.setId(Integer.parseInt(jTextID.getText()));
             p.setPreco(Double.parseDouble(jTextPreco.getText()));
             p.setDescricao(jTextDescricao.getText());
-            p.setQuantidade(Integer.parseInt(jTextQuantidade.getText()));
+            p.setQuantidade(Double.parseDouble(jTextQuantidade.getText()));
             ProdutoDao pd = new ProdutoDao();
             pd.cadastrar(p);
             JOptionPane.showMessageDialog(null,"Produto cadastrado com sucesso.","Sucesso",JOptionPane.INFORMATION_MESSAGE);
